drivers: haptics: Split CS40L5x header into private and public headers
Many structures and functions in the public CS40L5x do not need to be
exposed to userspace. Split the header accordingly into private
and public headers, and remove comments accordingly.
